requirements.txt
# ASTRA unified GUI and tools
customtkinter>=5.2
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
pandas>=2.0
pyarrow>=14.0
astropy>=6.0
skyfield>=1.46
pyserial>=3.5
sep>=1.2.0
scipy>=1.10
zwoasi>=0.2.0
